se cambian los valores de daño de int a float

+++ src/main/java/fiuba/algo3/algochess/model/juego/ProveedorConstantes.java
@@ -1,21 +1,21 @@
 package fiuba.algo3.algochess.model.juego;
 
 public class ProveedorConstantes {
-    public static int DANIO_ARMA_SOLDADO = 10;
-    public static int VIDA_SOLDADO = 100;
+    public static float DANIO_ARMA_SOLDADO = 10;
+    public static float VIDA_SOLDADO = 100;
     public static int COSTO_SOLDADO = 1;
-    public static int DANIO_DISTANCIA_SOLDADO = 0;
-    public static int DANIO_ARMA_JINETE = 5;
-    public static int VIDA_JINETE = 100;
+    public static float DANIO_DISTANCIA_SOLDADO = 0;
+    public static float DANIO_ARMA_JINETE = 5;
+    public static float VIDA_JINETE = 100;
     public static int COSTO_JINETE = 3;
-    public static int DANIO_DISTANCIA_JINETE = 15;
-    public static int CURACION_CURANDERO = 15;
-    public static int VIDA_CURANDERO = 75;
+    public static float DANIO_DISTANCIA_JINETE = 15;
+    public static float CURACION_CURANDERO = 15;
+    public static float VIDA_CURANDERO = 75;
     public static int COSTO_CURANDERO = 2;
-    public static int DANIO_ARMA_CATAPULTA = 0;
-    public static int VIDA_CATAPULTA = 50;
+    public static float DANIO_ARMA_CATAPULTA = 0;
+    public static float VIDA_CATAPULTA = 50;
     public static int COSTO_CATAPULTA = 5;
-    public static int DANIO_DISTANCIA_CATAPULTA = 20;
+    public static float DANIO_DISTANCIA_CATAPULTA = 20;
     public static int PUNTOS_INICIALES = 20;
     public static int TAMANIO_TABLERO = 20;
     public static int CANTIDAD_SOLDADOS_BATALLON = 3;
